pyproject.toml
[project]
name = "news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bs4>=0.0.2",
    "feedparser>=6.0.12",
    "gnews>=0.4.2",
    "huggingface-hub>=0.36.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
    "ruff>=0.14.6",
    "torch>=2.9.1",
    "transformers>=4.57.1",
]
